# Optimize multi-fragment unfiltering, part 2